backend/app/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.13.1
gitpython==3.1.40
detect-secrets==1.4.0
bandit==1.7.5
safety==2.3.5
pyyaml==6.0.1
langchain==0.0.350
openai==1.3.7